intermediate

+++ dsynk/src/main/java/com/csrg/dsynk/hub/controllers/v1/models/MessageDto.java
@@ -12,7 +12,7 @@ public class MessageDto
     String value;
 
     // specific to document
-    Integer line;
+    Double line;
 
     public RequestMessage asRequestMessage() {
